Fixing exception when there was no XDC file nor a template

+++ Hast.Abstractions/Hast.Xilinx.Abstractions/VivadoHardwareImplementationComposer.cs
@@ -66,7 +66,7 @@ namespace Hast.Xilinx.Abstractions
             {
                 File.Copy(xdcFilePath, xdcFileTemplatePath);
             }
-            else
+            else if (File.Exists(xdcFileTemplatePath))
             {
                 File.Copy(xdcFileTemplatePath, xdcFilePath, true);
             }
